Move swapByteOrderN into mastercore_convert.h/.cpp
