+++ src/mastercore_tx.cpp
@@ -47,6 +47,7 @@ using namespace leveldb;
 
 using namespace mastercore;
 
+#include "mastercore_convert.h"
 #include "mastercore_dex.h"
 #include "mastercore_tx.h"
 #include "mastercore_sp.h"
